+++ HISTORY.md
@@ -4,7 +4,7 @@ Release History
 0.1.9 (unreleased)
 ------------------
 
-- Nothing changed yet.
+- Bugfixes, fix black version to black==20.8b1.
 
 
 0.1.8 (2020-12-16)